+++ src/main/kotlin/blackjack/domain/GameMembers.kt
@@ -1,5 +1,7 @@
 package blackjack.domain
 
+import blackjack.domain.Participant.Dealer
+
 class GameMembers(private val participants: Participants, private val dealer: Dealer) {
     fun allPlayers(): Participants = Participants(participants, dealer)
