+++ src/main/kotlin/blackjack/domain/Participant.kt
@@ -31,4 +31,29 @@ sealed class Participant {
             return totalSum
         }
     }
+
+    class Dealer(
+        private val deck: Deck,
+        private val displayName: String = "Dealer",
+    ) : Player(displayName) {
+        fun deal(participants: Participants) {
+            repeat(NUMBER_OF_DEAL_CARD) {
+                dealOneCardToEachPlayer(participants)
+            }
+        }
+
+        private fun dealOneCardToEachPlayer(participants: Participants) {
+            participants.allPlayers().forEach {
+                it.receiveCard(deck.draw())
+            }
+        }
+
+        fun hit(player: Player) {
+            player.receiveCard(deck.draw())
+        }
+
+        companion object {
+            private const val NUMBER_OF_DEAL_CARD = 2
+        }
+    }
 }
