+++ src/test/kotlin/domain/GameTest.kt
@@ -2,11 +2,11 @@ package domain
 
 import blackjack.domain.Card
 import blackjack.domain.CardNumber
-import blackjack.domain.Dealer
 import blackjack.domain.Deck
 import blackjack.domain.Game
 import blackjack.domain.GameMembers
 import blackjack.domain.HitCommand
+import blackjack.domain.Participant.Dealer
 import blackjack.domain.Participant.Player
 import blackjack.domain.Participants
 import blackjack.domain.Result
